Added column for update button. For now only prints 'update' when pressed. Updated Balance label to display the total balance of the transactions

+++ finance-app/src/main/java/com/kletinich/Controllers/TransactionWindowController.java
@@ -17,8 +17,14 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.layout.HBox;
+import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
+import javafx.util.Callback;
 
 public class TransactionWindowController {
+
+    private double totalBalance;
 
     @FXML private Label balanceLabel;
 
@@ -35,16 +41,26 @@ public class TransactionWindowController {
     @FXML private TableColumn<Transaction, Timestamp> dateColumn;
     @FXML private TableColumn<Transaction, String> budgetSavingColumn;
     @FXML private TableColumn<Transaction, String> noteColumn;
-    @FXML private TableColumn<Transaction, Button> actionsColumn;
+    @FXML private TableColumn<Transaction, Button> deleteColumn;
+    @FXML private TableColumn<Transaction, Button> updateColumn;
 
     @FXML private Button newTransactionButton;
 
     @FXML
     public void initialize(){
+        totalBalance = 0;
+
         transactionsTable.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY);
 
         List<Transaction> transactionsList = TransactionDAO.getTransactions(null, null, null);
         ObservableList<Transaction> transactions = FXCollections.observableArrayList(transactionsList);
+
+        for(Transaction t: transactionsList){
+            totalBalance += t.getAmount();
+        }
+
+        balanceLabel.setText("Total balance: " + String.valueOf(totalBalance));
+
         idColumn.setCellValueFactory(new PropertyValueFactory<>("transactionID"));
         typeColumn.setCellValueFactory(new PropertyValueFactory<>("type"));
         amountColumn.setCellValueFactory(new PropertyValueFactory<>("amount"));
@@ -52,39 +68,81 @@ public class TransactionWindowController {
         dateColumn.setCellValueFactory(new PropertyValueFactory<>("date"));
         noteColumn.setCellValueFactory(new PropertyValueFactory<>("note"));
 
-        actionsColumn.setCellFactory(param -> new TableCell<>(){
+        // setting buttons in the deleteColumn for dynamic deletion
+        deleteColumn.setCellFactory(param -> new TableCell<>(){
             private final Button actionButton = new Button("Delete");
 
             {
-                actionButton.setOnAction(event -> {
-                    Transaction transaction = getTableView().getItems().get(getIndex());
-
-                    System.out.println("Deleting " + transaction);
-                    TransactionDAO.deleteTransactionByID(transaction.getTransactionID());
-                    getTableView().getItems().remove(transaction);
-                }
-                
-                );
+                actionButton.setOnAction(new EventHandler<ActionEvent>() {
+                    @Override
+                    public void handle(ActionEvent event) {
+                        deleteButtonPressed(getIndex());
+                    }
+                });
             }
         
         
-            @Override
-            protected void updateItem(Button btn, boolean isEmpty){
-                super.updateItem(btn, isEmpty);
+                @Override
+                protected void updateItem(Button btn, boolean isEmpty){
+                    super.updateItem(btn, isEmpty);
 
-                if(isEmpty){
-                    setGraphic(null);
-                }
+                    if(isEmpty){
+                        setGraphic(null);
+                    }
 
-                else{
-                    setGraphic(actionButton);
+                    else{
+                        setGraphic(actionButton);
+                    }
                 }
             }
-        }
+        );
+
+        // setting buttons in the updateColumn for dynamic update
+        updateColumn.setCellFactory(param -> new TableCell<>(){
+            private final Button actionButton = new Button("Update");
+
+            {
+                actionButton.setOnAction(new EventHandler<ActionEvent>() {
+                    @Override
+                    public void handle(ActionEvent event) {
+                        updateButtonPressed(getIndex());
+                    }
+                });
+            }
+        
+        
+                @Override
+                protected void updateItem(Button btn, boolean isEmpty){
+                    super.updateItem(btn, isEmpty);
+
+                    if(isEmpty){
+                        setGraphic(null);
+                    }
+
+                    else{
+                        setGraphic(actionButton);
+                    }
+                }
+            }
         );
 
 
         transactionsTable.setItems(transactions);
+    }
+
+    // when a delete button is pressed, delete the transaction associated with the index
+    public void deleteButtonPressed(int index){
+        if(index >= 0){
+            Transaction transaction = deleteColumn.getTableView().getItems().get(index);
+            System.out.println("Deleting " + transaction);
+            TransactionDAO.deleteTransactionByID(transaction.getTransactionID());
+            deleteColumn.getTableView().getItems().remove(transaction);
+        }
+    }
+
+    // when an update button is pressed, update the transaction associated with the index
+    public void updateButtonPressed(int index){
+        System.out.println("Update");
     }
     
 }
